antialiasing bricked the renders 8.2.

+++ src/vec3.h
@@ -1,24 +1,23 @@
 #ifndef VEC3_H
 #define VEC3_H
 
-#include <cmath>
-#include <iostream>
+#include "rtweekend.h"
 
 class vec3
 {
     public:
-        float e[3];
+        double e[3];
 
         vec3()  :   e{0, 0, 0} {}
-        vec3(float e0, float e1, float e2)  :   e{e0, e1, e2} {}
+        vec3(double e0, double e1, double e2)  :   e{e0, e1, e2} {}
         
-        float x() const {return e[0];}
-        float y() const {return e[1];}
-        float z() const {return e[2];}
+        double x() const {return e[0];}
+        double y() const {return e[1];}
+        double z() const {return e[2];}
 
         vec3 operator-() const {return vec3(-e[0], -e[1], -e[2]);}
-        float operator[](int i) const {return e[i];}
-        float& operator[](int i) {return e[i];}
+        double operator[](int i) const {return e[i];}
+        double& operator[](int i) {return e[i];}
 
         vec3& operator+=(const vec3& v)
         {
@@ -28,7 +27,7 @@ class vec3
             return *this;
         }
 
-        vec3& operator*=(float t)
+        vec3& operator*=(double t)
         {
             e[0] *= t;
             e[1] *= t;
@@ -36,17 +35,17 @@ class vec3
             return *this;
         }
 
-        vec3& operator/=(float t)
+        vec3& operator/=(double t)
         {
             return *this *= 1/t;
         }
 
-        float length() const
+        double length() const
         {
             return std::sqrt(lengthSquared());
         }
 
-        float lengthSquared() const 
+        double lengthSquared() const 
         {
             return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
         }
@@ -69,21 +68,21 @@ inline vec3 operator-(const vec3& u, const vec3& v)
     return vec3(u.e[0] - v.e[0], u.e[1] - v.e[1], u.e[2] - v.e[2]);
 }
 
-inline vec3 operator*(float t, const vec3& v)
+inline vec3 operator*(double t, const vec3& v)
 {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline vec3 operator*(const vec3& v, float t){
+inline vec3 operator*(const vec3& v, double t){
     return t * v;
 }
 
-inline vec3 operator/(const vec3& v, float t)
+inline vec3 operator/(const vec3& v, double t)
 {
     return (1/t) * v;
 }
 
-inline float dot(const vec3& u, const vec3& v)
+inline double dot(const vec3& u, const vec3& v)
 {
     return u.e[0] * v.e[0]
          + u.e[1] * v.e[1]
